# src/main/resources/openapi/SuperBlog-openapi.yml
openapi: 3.0.0
info:
  title: Super Blog Service API
  version: 0.0.1

paths:
  /super-blog/hello-world:
    get:
      summary: Print 'Hello World!'
      operationId: helloWorld
      responses:
        '200':
          description: Succesful operation
          content:
            text/plain:
              schema:
                type: string

  /super-blog:
    get:
      summary: Get the blog / List all blog's posts
      description: Get the blog / List all blog's posts
      operationId: getBlog
      responses:
        '200':
          description: Succesful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'


  /super-blog/post:
    post:
      summary: Create a new post
      description: Create a new post
      operationId: createPost
      requestBody:
        $ref: '#/components/requestBodies/PostBody'
      responses:
        '201':
          $ref: '#/components/responses/PostResponse'

  /super-blog/post/{postId}:
    get:
      summary: Get a post by ID
      description: Get a post by ID
      operationId: getPost
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '200':
          $ref: '#/components/responses/PostResponse'

    put:
      summary: Update a post by ID
      description: Update a post by ID
      operationId: updatePost
      parameters:
        - $ref: '#/components/parameters/postId'
      requestBody:
        $ref: '#/components/requestBodies/PostBody'
      responses:
        '200':
          $ref: '#/components/responses/PostResponse'

    delete:
      summary: Delete a post by ID
      description: Delete a post by ID
      operationId: deletePost
      parameters:
        - $ref: '#/components/parameters/postId'
      responses:
        '204':
          description: Succesful operation

components:
  parameters:
    postId:
      in: path
      name: postId
      schema:
        type: string
        minLength: 1
      required: true
      description: String ID of the post to get

  requestBodies:
    PostBody:
      description: Create a new post
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
      required: true

  responses:
    PostResponse:
      description: Succesful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'

  schemas:
    Post:
      type: object
      properties:
        postId:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 0
          maxLength: 100
        content:
          type: string
          minLength: 1
          maxLength: 700
        dateCreated:
          type: string
          format: date-time
        dateModified:
          type: string
          format: date-time
      required:
        - content